Fix numerous compiler warnings.

+++ cpu/rl78/rl78.h
@@ -1,6 +1,8 @@
 #ifndef RL78_H
 #define RL78_H
 
+#include <stdint.h>
+
 #include "sfrs.h"
 #include "sfrs-ext.h"
 
@@ -8,7 +10,7 @@
 #define CLOCK_CHANNEL 0
 #define CLOCK_SCALER  15       // Use f_CLK / 2^15.
 
-typedef unsigned short clock_time_t;
+typedef uint32_t clock_time_t;
 typedef unsigned short uip_stats_t;
 
 #endif // RL78_H
